+++ Packages/com.mimylab.fukuroudon/Runtime/ActiveRelay/Scripts/ActiveRelayByPlayerRespawn.cs
@@ -1,0 +1,33 @@
+﻿/*
+Copyright (c) 2024 Mimy Quality
+Released under the MIT license
+https://opensource.org/licenses/mit-license.php
+*/
+
+namespace MimyLab
+{
+    using UdonSharp;
+    using UnityEngine;
+    using VRC.SDKBase;
+    //using VRC.Udon;
+
+    [AddComponentMenu("Fukuro Udon/Active Relay/ActiveRelay by Player Respawn")]
+    [UdonBehaviourSyncMode(BehaviourSyncMode.None)]
+    public class ActiveRelayByPlayerRespawn : ActiveRelayBy
+    {
+        [SerializeField]
+        private bool _localOnly = true;
+
+        private void Reset()
+        {
+            _actionType = ActiveRelayActivateType.Activate;
+        }
+
+        public override void OnPlayerRespawn(VRCPlayerApi player)
+        {
+            if (_localOnly && !player.isLocal) { return; }
+
+            DoAction();
+        }
+    }
+}